requirements.txt
# ── LangChain / LangGraph ─────────────────────────────────────────────────────
langchain>=0.2.16,<0.3
langchain-community>=0.2.16,<0.3
langchain-openai>=0.1.23,<0.2
langchain-core>=0.2.38,<0.3
langgraph>=0.2.14,<0.3

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── PDF processing ────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0
pytesseract>=0.3.13
Pillow>=10.4.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.112.0,<0.115
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=24.1.0

# ── Config & logging ──────────────────────────────────────────────────────────
pydantic>=2.8.0
pydantic-settings>=2.4.0
structlog>=24.4.0

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=1.40.0
